Rushil/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
accelerate>=0.24.0
ultralytics>=8.0.0
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.0.0
numpy>=1.24.0
asyncio-mqtt>=0.13.0
psutil>=5.9.0
aiofiles>=23.2.0
pydantic>=2.5.0
requests>=2.31.0
hf_transfer
aiohttp